Namespace is an array of strings. Output property returns an array of outputs per namespace.

+++ src/server/api/routers/get-token.ts
@@ -5,14 +5,24 @@ import _get from "lodash/get";
 import { type DesignToken } from "style-dictionary/types/DesignToken";
 import { buildFile } from "~/utils/build-file";
 import { removeFile } from "~/utils/remove-file";
-import { getResponse } from "~/utils/get-response";
+import { getErrorOutput, getSuccessOutput } from "~/utils/get-output";
+
+type Output = {
+  id: string;
+  output: string;
+};
+
+type Response = {
+  id: string;
+  output: Output[];
+};
 
 export const getToken = createTRPCRouter({
   get: publicProcedure
     .input(
       z.object({
         id: z.string(),
-        namespace: z.string(),
+        namespace: z.array(z.string()),
         // transform: z.enum(['scss']),
       })
     )
@@ -24,27 +34,39 @@ export const getToken = createTRPCRouter({
           },
         });
         if (row) {
-          const objectByNamespace = _get(
-            JSON.parse(row.file),
-            input.namespace
-          ) as DesignToken;
-          if (objectByNamespace) {
-            const outputDirectory = buildFile({
-              token: objectByNamespace,
-              id: input.id,
-            });
-            const response = getResponse({
-              path: outputDirectory,
-              namespace: input.namespace,
-              id: input.id,
-            });
-            removeFile(outputDirectory);
-            return response;
-          }
+          const response: Response = {
+            id: input.id,
+            output: [],
+          };
 
-          return JSON.stringify({
-            error: `No token for namespace ${input.namespace} was found.`,
-          });
+          for (const namespace of input.namespace) {
+            const objectByNamespace = _get(
+              JSON.parse(row.file),
+              namespace
+            ) as DesignToken;
+            if (objectByNamespace) {
+              const outputDirectory = buildFile({
+                token: objectByNamespace,
+                id: input.id,
+              });
+              const successOutput = getSuccessOutput({
+                path: outputDirectory,
+                namespace: namespace,
+                id: input.id,
+              });
+              if (successOutput) {
+                removeFile(outputDirectory);
+                response.output.push(successOutput);
+              } else {
+                const errorOutput = getErrorOutput({
+                  namespace: namespace,
+                  id: input.id,
+                });
+                response.output.push(errorOutput);
+              }
+            }
+          }
+          return response;
         }
       }
       return null;
